# lib/units/api/swagger/api_v1.yaml
...
swagger: "2.0"
info:
  version: "2.4.0"
  title: Smartphone Test Farm
  description: Control and manages real Smartphone devices from browser and restful apis
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: STF Team
    email: contact@openstf.io
    url: http://openstf.io/
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
produces:
  - application/json
...
paths:
  /groups:
    x-swagger-router-controller: groups
    get:
      summary: Gets groups
      description: Returns the groups to which you belong
      operationId: getGroups
      tags:
        - groups
      parameters:
        - name: fields
          in: query
          description: Comma-seperated list of fields; only listed fields will be returned in response
          required: false
          type: string
        - name: owner
          in: query
          description: Selects the groups for which you are the owner (true) or a simple member (false); note that by not providing this parameter, it means all groups to which you belong are selected
          required: false
          type: boolean
      responses:
        "200":
          description: Groups information
          schema:
            $ref: "#/definitions/GroupListResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    delete:
      summary: Removes groups
      description: Removes the groups owned by you
      operationId: deleteGroups
      tags:
        - groups
      parameters:
        - name: groups
          in: body
          description: Groups to remove as a comma-separated list of group identifiers; note that by not providing this parameter it means all groups owned by you are removed
          required: false
          schema:
            $ref: "#/definitions/GroupsPayload"
      responses:
        "200":
          description: Groups removing is OK (or no groups to remove)
          schema:
            $ref: "#/definitions/Response"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => a device is currently booked or unremovable built-in group
              * 404: Not Found => unknown groups
              * 500: Internal Server Error
              * 503: Service Unavailable => server too busy or a lock on a resource is pending
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    post:
      summary: Creates a group
      description: Creates a group with you as owner
      operationId: createGroup
      tags:
        - groups
      parameters:
        - name: group
          in: body
          description: Group properties; at least one property is required
          required: true
          schema:
            $ref: "#/definitions/GroupPayload"
      responses:
        "201":
          description: Group information
          schema:
            $ref: "#/definitions/GroupResponse"
        default:
          description: >
            Unexpected Error:
              * 400: Bad Request => invalid format or semantic of properties
              * 401: Unauthorized => bad credentials
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
...
  /groups/{id}/devices:
    x-swagger-router-controller: groups
    get:
      summary: Gets the devices of a group
      description: Returns the devices of the group to which you belong
      operationId: getGroupDevices
      tags:
        - groups
      parameters:
        - name: id
          in: path
          description: Group identifier
          required: true
          type: string
        - name: bookable
          in: query
          description: Selects devices which could be potentially booked by that transient group (true => irrelevant for an origin group!), or selects all devices of the group (false); note that by not providing this parameter all devices of the group are selected
          type: boolean
          default: false
        - name: fields
          in: query
          description: Comma-separated list of device fields; only listed fields will be returned in response
          required: false
          type: string
      responses:
        "200":
          description: Group devices information
          schema:
            $ref: "#/definitions/DeviceListResponse"
        default:
          description: >
            Unexpected Error:
              * 400: Bad request => group is not transient
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown group
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    put:
      summary: Adds devices into a transient group
      description: Adds devices into a transient group owned by you
      operationId: addGroupDevices
      tags:
        - groups
      parameters:
        - name: id
          in: path
          description: Group identifier
          required: true
          type: string
        - name: devices
          in: body
          description: Devices to add as a comma-separated list of serials; note that by not providing this parameter it means all devices which could be potentially booked by that transient group are added into the latter
          required: false
          schema:
            $ref: "#/definitions/DevicesPayload"
      responses:
        "200":
          description: Group information (an empty group is returned if no change is made)
          schema:
            $ref: "#/definitions/GroupResponse"
        "409":
          description: Conflicts information
          schema:
            $ref: "#/definitions/ConflictsResponse"
        default:
          description: >
            Unexpected Error:
              * 400: Bad Request => group is not transient
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => quota is reached
              * 404: Not Found => unknown group or devices
              * 500: Internal Server Error
              * 503: Service Unavailable => server too busy or a lock on a resource is pending
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    delete:
      summary: Removes devices from a transient group
      description: Removes devices from a transient group owned by you
      operationId: removeGroupDevices
      tags:
        - groups
      parameters:
        - name: id
          in: path
          description: Group identifier
          required: true
          type: string
        - name: devices
          in: body
          description: Devices to remove as a comma-separated list of serials; note that by not providing this parameter it means all devices of the group are removed
          required: false
          schema:
            $ref: "#/definitions/DevicesPayload"
      responses:
        "200":
          description: Group information (an empty group is returned if no change is made)
          schema:
            $ref: "#/definitions/GroupResponse"
        default:
          description: >
            Unexpected Error:
              * 400: Bad Request => group is not transient
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown group or devices
              * 500: Internal Server Error
              * 503: Service Unavailable => server too busy or a lock on a resource is pending
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
...
  /user/devices:
    x-swagger-router-controller: user
    get:
      summary: User Devices
      description: The User Devices endpoint returns device list owner by current authorized user
      operationId: getUserDevices
      tags:
        - user
      parameters:
        - name: fields
          in: query
          description: Fields query parameter takes a comma seperated list of fields. Only listed field will be return in response
          required: false
          type: string
      responses:
        "200":
          description: Current User Devices List
          schema:
            $ref: "#/definitions/DeviceListResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    post:
      summary: Add a device to a user
      description: The User Devices endpoint will request stf server for a new device.
      operationId: addUserDevice
      tags:
        - user
      parameters:
        - name: device
          in: body
          description: Device to add
          required: true
          schema:
            $ref: "#/definitions/AddUserDevicePayload"
      responses:
        "200":
          description: Add User Device Status
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is already controlled or is not available
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
...
  /user/devices/{serial}/remoteConnect:
    x-swagger-router-controller: user
    post:
      summary: Remote Connect
      description: The device connect endpoint will request stf server to connect remotely
      operationId: remoteConnectUserDeviceBySerial
      tags:
        - user
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
      responses:
        "200":
          description: Remote Connect User Device Request Status
          schema:
            $ref: "#/definitions/RemoteConnectUserDeviceResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is not controlled by the user
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    delete:
      summary: Remote Disconnect
      description: The device connect endpoint will request stf server to disconnect remotely
      operationId: remoteDisconnectUserDeviceBySerial
      tags:
        - user
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
      responses:
        "200":
          description: Remote Disсonnect User Device Request Status
          schema:
            $ref: "#/definitions/Response"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is not controlled by the user
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
...
  /autotests:
    x-swagger-router-controller: autotests
    get:
      summary: Get devices for autotests run
      description: Create group and return serials of captured devices
      operationId: captureDevices
      tags:
        - autotests
      parameters:
        - name: timeout
          in: query
          description: timeout for group in seconds
          required: false
          type: integer
          minimum: 600
        - name: amount
          in: query
          description: Device amount needed for autotests run
          required: true
          type: integer
          minimum: 1
        - name: need_amount
          in: query
          description: need only specified amount, not less
          type: boolean
          required: false
        - name: abi
          in: query
          description: device abi
          type: string
          required: false
        - name: sdk
          in: query
          description: device sdk
          type: string
          required: false
        - name: model
          in: query
          description: device model
          type: string
          required: false
        - name: version
          in: query
          description: device os version
          type: string
          required: false
        - name: run
          in: query
          description: run identificator
          required: true
          type: string
      responses:
        "200":
          description: Group created for run
          schema:
            $ref: "#/definitions/GroupListResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown data
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
    delete:
      summary: delete autotests run data
      description: delete group and data created for autotests
      operationId: freeDevices
      tags:
        - autotests
      parameters:
        - name: group
          in: query
          description: Device amount needed for autotests run
          required: true
          type: string
      responses:
        "200":
          description: Group created for run
          schema:
            $ref: "#/definitions/GroupListResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown data
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
  /autotests/install/{serial}:
    x-swagger-router-controller: autotests
    post:
      summary: install apk on device by serial
      description: Installing apk to device from url
      operationId: installOnDevice
      tags:
        - autotests
      parameters:
        - name: serial
          in: path
          description: Device identifier (serial)
          required: true
          type: string
        - name: installProperties
          in: body
          schema:
            $ref: "#/definitions/AdbInstallFlagsPayload"
          description: List of flags for adb install
          required: true
      responses:
        "200":
          description: Remote Connect User Device Request Status
          schema:
            $ref: "#/definitions/RemoteConnectUserDeviceResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is not controlled by the user
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
  /autotests/useDevice:
    x-swagger-router-controller: autotests
    post:
      summary: Use device and remote controll it in one method
      description: Use device and remote controll it in one method
      operationId: useAndConnectDevice
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - autotests
      parameters:
        - name: serial
          in: body
          schema:
            type: object
            required:
              - serial
            properties:
              serial:
                type: string
                description: device serial
              title:
                type: string
                description: adb device serial
          description: serial of device, which used in action
          required: true
      responses:
        "200":
          description: stats added
          schema:
            $ref: "#/definitions/GroupListResponse"
        default:
          description: >
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown data
              * 500: Internal Server Error
          schema:
            $ref: "#/definitions/UnexpectedErrorResponse"
      security:
        - accessTokenAuth: [ ]
...
definitions:
  UnexpectedErrorResponse:
    required:
      - success
      - description
    properties:
      success:
        type: boolean
        default: false
      description:
        type: string
  Response:
    required:
      - success
      - description
    properties:
      success:
        type: boolean
        default: true
      description:
        type: string
  GroupResponse:
    required:
      - success
      - description
      - group
    properties:
      success:
        type: boolean
      description:
        type: string
      group:
        description: A null value means the group is unchanged
        type: object
  Conflict:
    type: object
    properties:
      devices:
        description: Devices in conflict
        type: array
        items:
          type: string
      date:
        description: Timeslot in conflict
        type: object
        properties:
          start:
            type: string
            format: date-time
          stop:
            type: string
            format: date-time
      group:
        description: Name of the group in conflict
        type: string
      owner:
        description: Owner of the group in conflict
        type: object
        properties:
          email:
            type: string
          name:
            type: string
  ConflictsResponse:
    required:
      - success
      - description
      - conflicts
    properties:
      success:
        type: boolean
      description:
        type: string
      conflicts:
        description: >
          List of conflicts with the current group operation:
           * adding a device into the group
           * updating the schedule of the group
        type: array
        items:
          $ref: '#/definitions/Conflict'
  GroupListResponse:
    required:
      - success
      - description
      - groups
    properties:
      success:
        type: boolean
      description:
        type: string
      groups:
        type: array
        items:
          type: object
...
  DeviceListResponse:
    required:
      - success
      - description
      - devices
    properties:
      success:
        type: boolean
      description:
        type: string
      devices:
        type: array
        items:
          type: object
...
  RemoteConnectUserDeviceResponse:
    required:
      - success
      - description
      - remoteConnectUrl
    properties:
      success:
        type: boolean
      description:
        type: string
      remoteConnectUrl:
        type: string
  AddUserDevicePayload:
    description: payload object for adding device to user
    required:
      - serial
    properties:
      serial:
        description: Device Serial
        type: string
      timeout:
        description: Device timeout in ms. If device is kept idle for this period, it will be automatically disconnected. Default is provider group timeout
        type: integer
  GroupPayload:
    description: Payload object for creating/updating a group
    properties:
      name:
        description: Group Name; default value => generated at runtime
        type: string
        pattern: '^[0-9a-zA-Z-_./: ]{1,50}$'
      startTime:
        description: Group starting time (in UTC, conforming to RFC 3339 section 5.6); default value => group creation time
        type: string
        format: date-time
      stopTime:
        description: Group expiration time (in UTC, conforming to RFC 3339 section 5.6); default value => startTime + 1 hour
        type: string
        format: date-time
      class:
        description: Group class; privileged value => debug, bookable, standard
        type: string
        enum:
          - once
          - bookable
          - hourly
          - daily
          - weekly
          - monthly
          - quaterly
          - halfyearly
          - yearly
          - debug
          - standard
        default: once
      repetitions:
        description: Group repetitions; default value => 0
        type: integer
        minimum: 0
      state:
        description: Group state; default value => pending or ready for bookable/standard classes
        type: string
        enum:
          - pending
          - ready
  GroupsPayload:
    description: Payload object for adding/removing groups
    properties:
      ids:
        description: Comma-separated list of identifiers
        type: string
...
  DevicesPayload:
    description: Payload object for adding/removing devices
    properties:
      serials:
        description: Comma-separated list of serials
        type: string
...
  AdbInstallFlagsPayload:
    description: List of flags for adb install
    required:
      - url
    properties:
      installFlags:
        type: array
        items:
          type: string
      url:
        type: string
securityDefinitions:
  accessTokenAuth:
    type: apiKey
    name: authorization
    in: header